Refactor code to improve readability and maintainability
- Removed unnecessary static_casts in various files to simplify expressions and enhance clarity.
- Added [[maybe_unused]] attribute to unused parameters and functions to suppress compiler warnings.
- Cleaned up constructor and destructor definitions in LLM-related classes to adhere to header-only implementation.
- Improved handling of empty containers in calculations to prevent potential division by zero errors.
- Streamlined logging and debug statements for better traceability.
- Updated comments for clarity and consistency across the codebase.

+++ src/llm/lora_router.cpp
@@ -533,7 +533,7 @@ RoutingDecision LoRARouter::selectByABTest(
     float cumulative = 0.0f;
     
     std::string selected_adapter = {};
-    for (size_t i = 0; i < ab_test_config_-> static_cast<int>(adapter_ids.size()); ++i) {
+    for (size_t i = 0; i < ab_test_config_->adapter_ids.size(); ++i) {
         cumulative += ab_test_config_->traffic_splits[i];
         if (rand_val <= cumulative) {
             selected_adapter = ab_test_config_->adapter_ids[i];
